+++ mne/io/meas_info.py
@@ -1716,12 +1716,13 @@ def _merge_info(infos, force_update_to_first=False, verbose=None):
                     'comps', 'custom_ref_applied', 'description',
                     'experimenter', 'file_id', 'highpass',
                     'hpi_subsystem', 'events',
-                    'line_freq', 'lowpass', 'meas_date', 'meas_id',
+                    'line_freq', 'lowpass', 'meas_id',
                     'proj_id', 'proj_name', 'projs', 'sfreq', 'gantry_angle',
                     'subject_info', 'sfreq', 'xplotter_layout', 'proc_history']
     for k in other_fields:
         info[k] = _merge_info_values(infos, k)
 
+    info['meas_date'] = infos[0]['meas_date']
     info._check_consistency()
     return info